adapters/python-libipld/requirements.txt
python-libipld==0.1.1